# full_node/discovery/src/openapi.json
{"openapi": "3.1.0", "info": {"title": "Full Node - Discovery", "version": "0.1.0"}, "paths": {"/authentication/authenticate": {"post": {"tags": ["authentication"], "summary": "Authenticate", "operationId": "authenticate_authentication_authenticate_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthenticationRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthenticationResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/authentication/refresh_access": {"post": {"tags": ["authentication"], "summary": "Refresh Access", "operationId": "refresh_access_authentication_refresh_access_post", "parameters": [{"required": true, "schema": {"type": "string", "title": "Refresh Token"}, "name": "refresh_token", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/offloading/offload_task_and_get_result": {"post": {"tags": ["offloading"], "summary": "Offload Task And Get Result", "operationId": "offload_task_and_get_result_offloading_offload_task_and_get_result_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/OffloadRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OffloadResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/offloading/offload_task": {"post": {"tags": ["offloading"], "summary": "Offload Task", "operationId": "offload_task_offloading_offload_task_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/OffloadRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OffloadResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/offloading/poll_result": {"post": {"tags": ["offloading"], "summary": "Poll Result", "operationId": "poll_result_offloading_poll_result_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PollResultRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OffloadResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"AuthenticationRequest": {"properties": {"did": {"type": "string", "title": "Did", "example": "did:meca:0x0fa21fd3d11d2cd5e6cdef2c7cd6531a25a5964f"}, "credential": {"type": "object", "title": "Credential", "example": {"credential": {"context": "https://www.w3.org/2018/credentials/v1", "id": "234508ed-fc5b-4ea4-88ac-f04bc657f470", "cptId": 2000000, "issuer": "did:meca:0x52c328ef8b382b1d71cc262b868d803a137ab8d8", "issuanceDate": 1680058078, "expirationDate": 4833658078, "claim": {"gender": "M", "name": "Chai", "DID": "did:meca:0x0fa21fd3d11d2cd5e6cdef2c7cd6531a25a5964f", "age": 29}, "proof": {"creator": "did:meca:0x52c328ef8b382b1d71cc262b868d803a137ab8d8", "salt": {"gender": "ZZijO", "name": "6fLwp", "DID": "cXQ86", "age": "FjOSj"}, "created": 1680058078, "type": "Secp256k1", "signatureValue": "zNDJ6NiFvu0yoFYBx7V3ZZVPaBPEMvqBieHbdhUFb8t3YWzRvje5dm8LM8p+OJ85Gq26u+yITTCkGwQlzUVKZxs="}, "type": ["VerifiableCredential", "original"], "signature": "zNDJ6NiFvu0yoFYBx7V3ZZVPaBPEMvqBieHbdhUFb8t3YWzRvje5dm8LM8p+OJ85Gq26u+yITTCkGwQlzUVKZxs=", "hash": "0x066d5c0edf9808c11ce4a45f5679b9cba9dacbbc99800f362063887d8ac83ca6", "signatureThumbprint": "{\"claim\":{\"DID\":\"did:meca:0x0fa21fd3d11d2cd5e6cdef2c7cd6531a25a5964f\",\"age\":29,\"gender\":\"M\",\"name\":\"Chai\"},\"context\":\"https://www.w3.org/2018/credentials/v1\",\"cptId\":2000000,\"expirationDate\":4833658078,\"id\":\"234508ed-fc5b-4ea4-88ac-f04bc657f470\",\"issuanceDate\":1680058078,\"issuer\":\"did:meca:0x52c328ef8b382b1d71cc262b868d803a137ab8d8\",\"type\":[\"VerifiableCredential\",\"original\"]}", "salt": {"gender": "ZZijO", "name": "6fLwp", "DID": "cXQ86", "age": "FjOSj"}, "credentialType": "ORIGINAL", "proofType": "Secp256k1"}}}}, "type": "object", "required": ["did", "credential"], "title": "AuthenticationRequest"}, "AuthenticationResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token", "description": "JWT format token"}, "access_token_type": {"type": "string", "title": "Access Token Type", "example": "Bearer"}, "refresh_token": {"type": "string", "title": "Refresh Token", "description": "JWT format token"}, "refresh_token_type": {"type": "string", "title": "Refresh Token Type", "example": "Bearer"}}, "type": "object", "required": ["access_token", "access_token_type", "refresh_token", "refresh_token_type"], "title": "AuthenticationResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "OffloadRequest": {"properties": {"did": {"type": "string", "title": "Did", "example": "did:meca:0x0fa21fd3d11d2cd5e6cdef2c7cd6531a25a5964f"}, "task_id": {"type": "string", "title": "Task Id", "example": "001"}, "container_reference": {"type": "string", "title": "Container Reference", "example": "sampleserver:latest"}, "content": {"type": "string", "title": "Content", "example": "{\"name\": \"meca dev\"}"}, "resource": {"allOf": [{"$ref": "#/components/schemas/Resources"}], "title": "Resource", "example": "{\"cpu\": 2, \"memory\": 256}"}, "runtime": {"type": "string", "title": "Runtime", "example": "microVM"}, "use_gpu": {"type": "boolean", "title": "Use Gpu", "default": false, "example": true}, "gpu_count": {"type": "integer", "title": "Gpu Count", "example": 1}}, "type": "object", "required": ["did", "task_id", "container_reference", "content"], "title": "OffloadRequest"}, "OffloadResponse": {"properties": {"transaction_id": {"type": "string", "title": "Transaction Id", "example": "001"}, "task_id": {"type": "string", "title": "Task Id", "example": "001"}, "status": {"type": "integer", "title": "Status", "example": 1}, "response": {"type": "string", "title": "Response", "example": "Hello World"}, "error": {"type": "string", "title": "Error", "example": ""}}, "type": "object", "required": ["transaction_id", "task_id", "status", "response"], "title": "OffloadResponse"}, "PollResultRequest": {"properties": {"did": {"type": "string", "title": "Did", "example": "did:meca:0x0fa21fd3d11d2cd5e6cdef2c7cd6531a25a5964f"}, "transaction_id": {"type": "string", "title": "Transaction Id"}}, "type": "object", "required": ["did", "transaction_id"], "title": "PollResultRequest"}, "Resources": {"properties": {"cpu": {"type": "integer", "exclusiveMinimum": 0.0, "title": "Cpu", "default": 1, "example": 2}, "memory": {"type": "integer", "exclusiveMinimum": 0.0, "title": "Memory", "default": 128, "example": 256}}, "type": "object", "title": "Resources"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}